ref(dynamic-sampling): dedupe schema, use SubmitButton
Extract shared targetSampleRateSchema to a single utils file and
replace raw Button type=submit with form.SubmitButton which handles
priority, busy state, and submission automatically.
